Edit departments
* work

* basic edits

* basic edits

* donzo

+++ src/main/java/com/viktor/timeofftests/models/Department.java
@@ -12,7 +12,7 @@ import java.util.Objects;
 public class Department {
     private int id;
     private String name;
-    private int allowance;
+    private double allowance;
     private boolean includePublicHolidays;
     private boolean isAccuredAllowance;
     private int companyId;
